Backend code for new videos search system (#838)
Created a MySQL table to index our videos rather than Solr.  To do
this we need to create a FULLTEXT index, and that in turn requires a
MyISAM table.  So there's a bunch of code to hack around that.

Added code to update the index table when needed.

Moved the setup_indexes management command to the util package.  I
think it makes more sense here because it sets up indexes for other
apps than just videos.

Made drop_all_tables set the collation to utf8_bin.  This is how
things are on production, so I think we should match it on dev.