Switched team videos pages to use the new search code (#838)
Took this opportunity to add filters for the new pages based on
completed languages.  Added some Video queryset methods to handle the
SQL behind this.

Reworked the old team videos pages a bunch now that they use actual
TeamVideo objects instead of search records.

Removed a bunch of teams tests that weren't working because they
relied on the hitting the search page, but that doesn't work in the
tests.  Honestly, I don't think these tests are very useful.  They
were definitely testing more than 1 thing, so when they failed it was
hard to understand what caused the failure.  Also, many of the things
they were testing was legacy teams code.